# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  furniture-commission_internal_service.CreatePurchaseItemRequest:
    properties:
      product_name:
        example: 真皮沙发
        type: string
      purchase_price:
        type: number
      quantity:
        minimum: 1
        type: integer
      sku_id:
        type: integer
      sku_name:
        example: 真皮沙发-棕色-三座
        type: string
    required:
    - purchase_price
    - quantity
    - sku_id
    type: object
  furniture-commission_internal_service.CreatePurchaseOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/furniture-commission_internal_service.CreatePurchaseItemRequest'
        minItems: 1
        type: array
      remark:
        example: 紧急采购
        type: string
      store_id:
        type: integer
      supplier_id:
        type: integer
      supplier_name:
        example: 某某供应商
        type: string
    required:
    - items
    type: object
...
  internal_handler.ConfirmReceiptRequest:
    properties:
      remark:
        example: 入库备注
        type: string
      warehouse_id:
        example: 1
        type: integer
    type: object
...
  internal_handler.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
      timestamp:
        example: 1715000000
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: 家具提成管理系统后端 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 家具提成管理系统 API
  version: "1.0"
paths:
...
  /inventory/stock:
    get:
      consumes:
      - application/json
      description: 根据仓库ID和SKU ID查询库存信息
      parameters:
      - description: 仓库ID
        format: int64
        in: query
        name: warehouse_id
        required: true
        type: integer
      - description: SKU ID
        format: int64
        in: query
        name: sku_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 查询库存
      tags:
      - 库存管理
  /inventory/stocks:
    get:
      consumes:
      - application/json
      description: 分页查询仓库库存列表，支持按仓库和关键词筛选
      parameters:
      - description: 仓库ID，不传或传0则查询所有仓库
        format: int64
        in: query
        name: warehouse_id
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取库存列表
      tags:
      - 库存管理
  /inventory/transactions:
    get:
      consumes:
      - application/json
      description: 分页查询库存流水列表，支持按仓库、类型、日期范围筛选
      parameters:
      - description: 仓库ID，不传或传0则查询所有仓库
        in: query
        name: warehouse_id
        type: integer
      - description: 流水类型
        in: query
        name: type
        type: integer
      - description: 开始日期
        in: query
        name: start_date
        type: string
      - description: 结束日期
        in: query
        name: end_date
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取库存流水
      tags:
      - 库存管理
...
  /purchases:
    get:
      consumes:
      - application/json
      description: 分页查询采购订单列表，支持按门店、状态、关键词筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 门店ID
        format: int64
        in: query
        name: store_id
        type: integer
      - description: 订单状态
        format: int32
        in: query
        name: status
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取采购订单列表
      tags:
      - 采购管理
    post:
      consumes:
      - application/json
      description: 创建新的采购订单，需要管理员权限
      parameters:
      - description: 仓库ID
        format: int64
        in: query
        name: warehouse_id
        type: integer
      - description: 创建采购订单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.CreatePurchaseOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建采购订单
      tags:
      - 采购管理
  /purchases/{id}:
    get:
      consumes:
      - application/json
      description: 根据采购订单ID获取订单详细信息
      parameters:
      - description: 采购订单ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取采购订单详情
      tags:
      - 采购管理
  /purchases/{id}/approve:
    put:
      consumes:
      - application/json
      description: 审核通过采购订单，需要管理员权限
      parameters:
      - description: 采购订单ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 审核采购订单
      tags:
      - 采购管理
  /purchases/{id}/cancel:
    put:
      consumes:
      - application/json
      description: 取消采购订单，需要管理员权限
      parameters:
      - description: 采购订单ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 取消采购订单
      tags:
      - 采购管理
  /purchases/{id}/receipt:
    put:
      consumes:
      - application/json
      description: 确认采购订单入库，需要管理员权限
      parameters:
      - description: 采购订单ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 确认入库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ConfirmReceiptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 确认入库
      tags:
      - 采购管理
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token 认证，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"